refactor: rename userDataSyncProfilesStorageService to userDataProfileStorageService (#166397)
* refactor: rename userDataSyncProfilesStorageService to userDataProfileStorageService

* rename

* fix import
